migrated and rename more components
